==> Vehicle Dataset Analysis/1_Preprocess.ipynb <==
# 1_Preprocess

In this notebook, I will make a preliminary analysis of the dataset, checking for variable types, missing values, basic descriptive statistics and possible inconsistencies 

# Step 1 | Import libraries
Import the necessary libraries that we are going to use further on.

import pandas as pd
import function_utils as futils
pd.set_option('display.float_format', '{:.2f}'.format) # disable scientific notation
pd.set_option('display.max_rows', None) # display all rows

# Step 2 | Import Dataset
Import the dataset

# Read the Excel file into a DataFrame
data = pd.read_excel("BCA_EU_CaseStudy_MarketIntelligence.xlsx")

# Display the first few rows of the DataFrame
data.head()

# 1. Brief description

The dataset is composed of 1003  vehicles and 17 features (aka attributes or columns).

Columns description:<br>
*veichle_ID* : ID of each vehicle <br>
*selling_country*: The country that hosted the auction <br>
*buyer_country* : The country where the buyer is located <br>
*make* : : Brand of the vehicle’s manufacture <br>
*model* : Specific model within the make of the vehicle<br>
*mileage* : : Total distance the vehicle has traveled in kilometers <br>
*body_type* :  Category associated with the shape and structure of the vehicle (e.g., sedan, SUV, hatchback).<br>
*fuel_type* : The type of fuel the vehicle uses to operate (e.g., petrol, diesel).<br>
*colour* : : The exterior color of the vehicle. <br>
*registration_date* : Date when the vehicle was first registered<br>
*selling_date* : The date when the vehicle is sold.<br>
*selling_price* : : The price at which the vehicle is sold.<br>
*condition_grade* : A score reflecting the vehicles overall condition, based on physical and mechanical assessments.<br>
*conversion_rate* : Number of vehicles sold / Number vehicles that entered a sale since January 2024.<br>
*b2b_price* : B2B Benchmark value given by an external provider<br>
*b2c_price*: B2C Benchmark value given by an external provider. T<br>
*stock_turnover* : average number of days it takes for a vehicle with those characteristics (e.g., model, fuel
type, age, etc.) to be sold.<br>

print(data.info())

print("Unique values by feature:\n", data.nunique().sort_values())

# 2.  Preliminary adjustments

## 2.1. Missing values

rm_df_mis = futils.missing_reporter(data, True)
display(rm_df_mis)

We have missing values in variable *condition_grade* that we can replace by "missing" as a category

# Replaces missing values 
data["condition_grade"].fillna(value='missing', inplace=True)

display("After imputing:", futils.missing_reporter(data))

# 3. Explore the data

## 3.1 Categorical Data

# Selects only object (categorical) features 
categorical_features = data.select_dtypes(include=['object']).columns.to_list()

# Removes vehicleID since is an ID not interesting for exploration
categorical_features.remove("vehicle_ID")

# for every column in the categorical features print the name and value counts of each category
for col in categorical_features:
    print('\n Feature', col)
    print('\n', data[col].value_counts())

It doesn't appear that data has any problem, it could be the case that we had typos or inconsistent case usage (e.g., "BMW" vs "Bmw") or even misclassified or undefined categories but it's not the case

## 3.2 Numerical Data

# Selects only numerical features 
numerical_features = data.select_dtypes(include=['int','float']).columns

# Get summary statistics for numerical variables
data[numerical_features].describe() 

It doesn't appear that data has any problem, it could be the case that we had negative values in these features which wouldn't make sense, and could also be the case that conversion_rate had values greater than 100% but it's not. 
I can already have a sense of the data with this summary statistics and for example understand that selling prices are higher than the B2B benchmark on average but lower then the B2C. 

## 3.3 Date time variables 

### 3.3.1 selling_date 

# Extract year, month, and day components from the selling_date column
data['selling_year'] = data['selling_date'].dt.year
data['selling_month'] = data['selling_date'].dt.month
data['selling_day'] = data['selling_date'].dt.day

#### Year

# Check years present in the dataset
data.groupby(['selling_year'])['vehicle_ID'].count()

It does not make sense to have 2099 in selling year.
In a real world situation i would ask to change this typo, but since here i cannot do it, i will just remove this row

# Removes rows where year is greater than 2024
data = data[data['selling_year'] <= 2024]

# Check years present in the dataset after removing 2099
data.groupby(['selling_year'])['vehicle_ID'].count()

#### Month

# Check months present in the dataset
data.groupby(['selling_month'])['vehicle_ID'].count()

#### Day

# Check months present in the dataset
data.groupby(['selling_month', 'selling_day'])['vehicle_ID'].count()

### 3.3.2 registration_date     

# Extract year, month, and day components from the registration_date column
data['registration_year'] = data['registration_date'].dt.year
data['registration_month'] = data['registration_date'].dt.month
data['registration_day'] = data['registration_date'].dt.day

#### Year

# Check years present in the dataset
data.groupby(['registration_year'])['vehicle_ID'].count()

#### Month

# Check years present in the dataset
data.groupby(['registration_month'])['vehicle_ID'].count()

#### Day

# Check months present in the dataset
data.groupby(['registration_month', 'registration_day'])['vehicle_ID'].count()

## 3.4 Possible inconsistencies.
It is very frequent to find data inconsistencies: facts which do not make any sense in the context of a given problem and can bias your analysis. In this sub-section I will explore the data in depth in order to identify them.

- It wouldn't make sense to have vehicles with sales date before registration date

data[data['selling_date'] < data['registration_date']]

- Let's compare the vehicle's actual selling price (internal transaction data) to its benchmark B2B price (external reference data). This will helps measure how well a vehicle's selling price aligns with the market benchmark for similar vehicles. I'm not expecting that will arise high deviations

# Compute deviations
data['deviation_b2B'] = (data['selling_price'] - data['b2b_price']) / data['b2b_price']
data['deviation_b2C'] = (data['selling_price'] - data['b2c_price']) / data['b2c_price']

data[(data['deviation_b2B'] > 1) | (data['deviation_b2B'] < -1)].T

This vheicle has been sold for a lot more then the B2B benchmark price. Still I will leave it in the dataset

data[(data['deviation_b2C'] > 1) | (data['deviation_b2C'] < -1)].T

# 4. Export dataset

# Check columns and rows of dataset after all the changes
data.shape

# Export dataset as pkl to preserve the data types
data.to_pickle('data_preprocessed.pkl')
