==> Vehicle Dataset Analysis/2_Extract.ipynb <==
# 2_Extract

In this notebook I will create new variables and check for redundant variables that i can exclude 

# Import libraries
Import the necessary libraries that we are going to use further on.

import pandas as pd
import numpy as np
import function_utils as futils
pd.set_option('display.float_format', '{:.2f}'.format) # disable scientific notation
pd.set_option('display.max_rows', None) # display all rows

# Import Dataset
Import the  dataset

# Read the pickle file 'df_preprocessed.pkl' and store the data in a DataFrame called df
df = pd.read_pickle('data_preprocessed.pkl')

# Create a copy of the df DataFrame and assign it to preprocessed_df
preprocessed_df = df.copy()

# Display the first 5 rows of the case_study_df DataFrame
display(preprocessed_df.head(5))

# Check if data types were preserved
preprocessed_df.dtypes

# 1. Feature Extraction
In this section, some new features will be calculated based on the variables we already have

## 1.1 Age of the vehicle at the time of sale

Calculate the age of the vehicle at the time of sale by subtracting the registration year from the sale year. Vehicles in the leasing segment are typically newer. Calculating age helps confirm whether vehicles from this vendor align with this profile.

# Function to calculate precise month difference
def calculate_months_diff(start_date, end_date):
    return (end_date.year - start_date.year) * 12 + (end_date.month - start_date.month) + \
           np.sign((end_date.day - start_date.day)) * 0.5  # Adjust for partial months

# Apply the function 
preprocessed_df['vehicle_age_sale_time_months'] = preprocessed_df.apply(lambda x: calculate_months_diff(x['registration_date'], x['selling_date']), axis=1)


# Check the new variable summary stats
preprocessed_df['vehicle_age_sale_time_months'].describe() 

## 1.2 Price Deviation (B2B and B2C)

Already calculated in 1_Preprocess notebook

Evaluates how selling prices compare to market benchmarks. High positive deviations suggest strong pricing performance. Comparing deviations for the leasing segment vs. other segments highlights opportunities or risks in adopting the new vendor.

## 1.3  Price to mileage ratio

A high price-to-mileage ratio often indicates vehicles that retain value well. Leasing vehicles generally perform better here due to lower mileage.

preprocessed_df['price_to_mileage_ratio'] = preprocessed_df['selling_price'] / preprocessed_df['mileage']


## 1.4 Leasing Segment

Classify vehicles into leasing or non-leasing segments based on Age, Mileage and Model

### 1.4.1 Age

Here i will check the distribution of the vehicle's age by decile to see in which value i will set the treshold for considering a leasing car in the leasing segment

_, age_aggregated = futils.create_decile_segment(
    df=preprocessed_df,
    sort_column='vehicle_age_sale_time_months',
    segment_prefix='decile',
    n_deciles=10,
    ascending=True,
    agg_columns={'vehicle_age_sale_time_months': ['min', 'max', 'mean']}
)

age_aggregated

Age: Will consider leasing the ones with less than 36 months

### 1.4.2 Mileage

Here i will check the distribution of the vehicle's mileage by decile to see in which value i will set the treshold for considering a leasing car in the leasing segment

_, mileage_aggregated = futils.create_decile_segment(
    df=preprocessed_df,
    sort_column='mileage',
    segment_prefix='decile',
    n_deciles=10,
    ascending=True,
    agg_columns={'mileage': ['min', 'max', 'mean']}
)

mileage_aggregated

Milleage: Will consider leasing the ones with less than 45k kilometers

### 1.4.3 Models

Here i could create a model (classification model) to see  which numeric variable (price, conversion_rate, etc) best describes the `model` and use that variable to make the same analysis as before. But for time purposes, i will just assume that price is a good variable to describe the `model`  

_, selling_price_aggregated = futils.create_decile_segment(
    df=preprocessed_df,
    sort_column='selling_price',
    segment_prefix='decile',
    n_deciles=10,
    ascending=True,
    agg_columns={'selling_price': ['min', 'max', 'mean']}
)

selling_price_aggregated

Model: Will consider leasing the models with selling_price greater than 25k 

# Create 'leasing_segment' 
leasing_conditions = (
    (preprocessed_df['vehicle_age_sale_time_months'] < 36) &
    (preprocessed_df['mileage'] < 35000) &
    (preprocessed_df['selling_price'] > 25000)
)


preprocessed_df['leasing_segment'] = np.where(leasing_conditions, 'Leasing', 'Non-Leasing')


segment_counts = preprocessed_df['leasing_segment'].value_counts()
print("Segment Counts:\n", segment_counts)

#### For analysis purpose, i will now calculate new variables from the numeric ones creating intervals or categories that will help me build the presentation in a clear and easier way for non-technical people to  understand the graphs easier. For the choice of the intervals i will check the distribution by deciles and see which intervals are better

## 1.5 Mileage segment


# Define the bin edges
bins = [0, 5000, 7000, 10000,  np.inf]

# Define the labels for the segments
labels = ['<= 5.000 km', '5.001 - 7.000 km', '7.001 - 10.000 km', '>=10.001 km']

# Create a new column 'mileage_segment' based on the defined bins
preprocessed_df['mileage_segment'] = pd.cut(
    preprocessed_df['mileage'],
    bins=bins,
    labels=labels,
    right=True,  # Whether to include the right edge (default is True)
    include_lowest=True  # Whether the first interval should be left-inclusive
)

# Verify the segmentation
preprocessed_df['mileage_segment'].value_counts().sort_index()

## 1.6 Selling price


# Define the bin edges
bins = [0, 15000, 25000, 35000, 50000, np.inf]  # 5 edges for 4 segments

labels = ['<= 15.000 €', '15.001 - 25.000 €', '25.001 - 35.000 €', '35.001 - 50.000 €','>=50.001 €']

#Create a new column 'selling_price_segment' based on the defined bins
preprocessed_df['selling_price_segment'] = pd.cut(
    preprocessed_df['selling_price'],
    bins=bins,
    labels=labels,
    right=True,          # Includes the right edge in the bin
    include_lowest=True  # Includes the lowest value in the first bin
)

# Verify the segmentation
preprocessed_df['selling_price_segment'].value_counts().sort_index()


## 1.7 Conversion Rate

_, conversion_rate_aggregated = futils.create_decile_segment(
    df=preprocessed_df,
    sort_column='conversion_rate',
    segment_prefix='decile',
    n_deciles=10,
    ascending=True,
    agg_columns={'conversion_rate': ['min', 'max', 'mean']}
)

conversion_rate_aggregated

# Define the bin edges
bins = [0, 82, 88, 92, np.inf]  # 5 edges for 4 segments

labels = ['<= 82.00%', '82.01 - 88.00 %', '88.01 - 92.00 %', '>=92.01 %']

# Create a new column 'selling_price_segment' based on the defined bins
preprocessed_df['conversion_rate_segment'] = pd.cut(
    preprocessed_df['conversion_rate'],
    bins=bins,
    labels=labels,
    right=True,          # Includes the right edge in the bin
    include_lowest=True  # Includes the lowest value in the first bin
)

# Verify the segmentation
segment_counts = preprocessed_df['conversion_rate_segment'].value_counts().sort_index()
print(segment_counts)

## 1.8 B2B Price

_, b2b_price_aggregated = futils.create_decile_segment(
    df=preprocessed_df,
    sort_column='b2b_price',
    segment_prefix='decile',
    n_deciles=10,
    ascending=True,
    agg_columns={'b2b_price': ['min', 'max', 'mean']}
)

b2b_price_aggregated

# Define the bin edges
bins = [0, 15000, 25000, 35000, 50000, np.inf]  # 5 edges for 4 segments

labels = ['<= 15.000 €', '15.001 - 25.000 €', '25.001 - 35.000 €', '35.001 - 50.000 €','>=50.001 €']

# Create a new column 'selling_price_segment' based on the defined bins
preprocessed_df['b2b_price_segment'] = pd.cut(
    preprocessed_df['b2b_price'],
    bins=bins,
    labels=labels,
    right=True,          # Includes the right edge in the bin
    include_lowest=True  # Includes the lowest value in the first bin
)

# Verify the segmentation
segment_counts = preprocessed_df['b2b_price_segment'].value_counts().sort_index()
print(segment_counts)

## 1.9 B2C Price

_, b2c_price_aggregated = futils.create_decile_segment(
    df=preprocessed_df,
    sort_column='b2c_price',
    segment_prefix='decile',
    n_deciles=10,
    ascending=True,
    agg_columns={'b2c_price': ['min', 'max', 'mean']}
)

b2c_price_aggregated

# Define the bin edges
bins = [0, 15000, 25000, 35000, 50000, np.inf]  # 5 edges for 4 segments

labels = ['<= 15.000 €', '15.001 - 25.000 €', '25.001 - 35.000 €', '35.001 - 50.000 €','>=50.001 €']

# Create a new column 'selling_price_segment' based on the defined bins
preprocessed_df['b2c_price_segment'] = pd.cut(
    preprocessed_df['b2c_price'],
    bins=bins,
    labels=labels,
    right=True,          # Includes the right edge in the bin
    include_lowest=True  # Includes the lowest value in the first bin
)

# Verify the segmentation
segment_counts = preprocessed_df['b2c_price_segment'].value_counts().sort_index()
print(segment_counts)

## 1.9 Stock Turnover

_, stock_turnover_aggregated = futils.create_decile_segment(
    df=preprocessed_df,
    sort_column='stock_turnover',
    segment_prefix='decile',
    n_deciles=10,
    ascending=True,
    agg_columns={'stock_turnover': ['min', 'max', 'mean']}
)

stock_turnover_aggregated

# Define the bin edges
bins = [0, 40, 50, 60, np.inf]  # 5 edges for 4 segments

labels = ['<= 40 days', '41 - 50 days', '51 - 60 days', '>=61 days']

#Create a new column 'selling_price_segment' based on the defined bins
preprocessed_df['stock_turnover_segment'] = pd.cut(
    preprocessed_df['stock_turnover'],
    bins=bins,
    labels=labels,
    right=True,          # Includes the right edge in the bin
    include_lowest=True  # Includes the lowest value in the first bin
)

#Verify the segmentation
segment_counts = preprocessed_df['stock_turnover_segment'].value_counts().sort_index()
print(segment_counts)

## Choosing variables for analysis

columns_to_drop = [
    'mileage',
    'vehicle_age_sale_time_months',
    'decile_stock_turnover',
    'decile_selling_price'
]

# Identify and drop existing columns
existing_columns_to_drop = [col for col in columns_to_drop if col in preprocessed_df.columns]
df_segmented = preprocessed_df.drop(columns=existing_columns_to_drop)


# Check  data types
df_segmented.dtypes

# Export dataset as pkl to preserve the data types
df_segmented.to_pickle('data_extract.pkl')
