==> Vehicle Dataset Analysis/3_Insights.ipynb <==
# 3_Insights

In this notebook, I will make all the calculations needed to take the relevant isights for the presentation

# Import libraries
Import the necessary libraries that we are going to use further on.

import pandas as pd
import function_utils as futils

# Import Dataset
Import the dataset from previous analysis

# Read the pickle file 'df_extract.pkl' and store the data in a DataFrame called df
df = pd.read_pickle('data_extract.pkl')

# Create a copy of the df DataFrame and assign it to extract_df
extract_df = df.copy()

# Display the first 5 rows of the case_study_df DataFrame
display(extract_df.head(5))

# Check basic statistics of all dataset after extraction
extract_df.describe()

# 1. Caracterization of leasing vs non-leasing segment

# create quarter to analyze variables by quarter
extract_df['selling_quarter'] = extract_df['selling_date'].dt.to_period('Q')

# seting output file for analysis
output_file = 'leasing_segments_analysis.xlsx' 

# set the group variables i want to analyze and put on the output file for analysis
grouping_configurations = [
    (['selling_quarter', 'leasing_segment'], 'Quarterly Count'),
    (['selling_country', 'leasing_segment'], 'Selling Country'),
    (['buyer_country', 'leasing_segment'], 'Buyer Country'),
    (['make', 'leasing_segment'], 'Make'),
    (['make', 'model', 'leasing_segment'], 'Make Model'),
    (['body_type', 'leasing_segment'], 'Body Type'),
    (['fuel_type', 'leasing_segment'], 'Fuel Type'),
    (['colour', 'leasing_segment'], 'Colour Type'),
    (['condition_grade', 'leasing_segment'], 'Condition Grade'),
    (['mileage_segment', 'leasing_segment'], 'Mileage Segment'),
    (['selling_price_segment', 'leasing_segment'], 'Selling Price Segment'),
    (['conversion_rate_segment', 'leasing_segment'], 'Conversion Rate Segment'),
    (['b2b_price_segment', 'leasing_segment'], 'B2B Price Segment'),
    (['b2c_price_segment', 'leasing_segment'], 'B2C Price Segment'),
    (['stock_turnover_segment', 'leasing_segment'], 'Stock Turnover Segment')
]



# Running the function to group and add to excel for every pair of variables i want to group. 
# The function counts the number of veichles by grouping variables
for config in grouping_configurations:
    group_columns, sheet_name = config  # Unpack the tuple
    futils.group_and_write_to_excel(extract_df, group_columns, output_file, sheet_name)
